requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
cryptography>=3.4.8

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
coverage==7.3.2

# Development
python-decouple==3.8